Fixes issues introduced with multi-type support in subproperties.rb.
It was hardcoded to fetch only the first item of the list (when a list
happened to be provided). Now fetching whole list of types.

The warning issued when type was incorrect or invalid syntax, when
multiple types were specified, had incorrect ruby syntax (mistook the
'types' variable for the 'typelist' one, no '.join()' call was necessary
at all).