Raises exception on error and ignore some 'html tags'.
Now raising exception when html tag is either not opened or not closed.

Ignores any 'if' or 'debug' (case sensitive) "html tag" (they are misc
tags used in Ext JS to control code output). As the closing tags are not
in the same comment block (they're rather inline comments, which could
actually be ignored by jsDuck instead).

For now, if no --ignore-html is provided, these will automagically be
considered to ignore.