Adds multi tag matching support for boolean and improves Locale tag.
Now an array can be specified with valid matches for the @locale tag,
supporting also @Locale.

The base boolean tag handler now supports a list of matches to assign to
that tag, and uses the first one to register the tag as (so leave the
first as the lowercase one).